+++ src/main/java/cart/repository/CartItemRepository.java
@@ -6,11 +6,11 @@ import cart.domain.CartItem;
 import cart.domain.Member;
 import cart.domain.Product;
 import cart.exception.CartItemException;
-import cart.repository.dao.CartItemDao;
-import cart.repository.entity.CartItemEntity;
-import cart.repository.entity.CartItemWithMemberAndProductEntity;
-import cart.repository.entity.MemberEntity;
-import cart.repository.entity.ProductEntity;
+import cart.dao.CartItemDao;
+import cart.dao.entity.CartItemEntity;
+import cart.dao.entity.CartItemWithMemberAndProductEntity;
+import cart.dao.entity.MemberEntity;
+import cart.dao.entity.ProductEntity;
 import java.util.List;
 import java.util.stream.Collectors;
 import org.springframework.stereotype.Repository;
