+++ src/main/java/cart/repository/ProductRepository.java
@@ -4,8 +4,8 @@ import static cart.exception.ErrorMessage.NOT_FOUND_PRODUCT;
 
 import cart.domain.Product;
 import cart.exception.ProductException;
-import cart.repository.dao.ProductDao;
-import cart.repository.entity.ProductEntity;
+import cart.dao.ProductDao;
+import cart.dao.entity.ProductEntity;
 import java.util.List;
 import java.util.stream.Collectors;
 import org.springframework.stereotype.Repository;
